[PC] Blazing Racer
Blazing Racer [PC] is a 2D Multiplayer Racing Game allowing up-to 2
players to compete each other in a challenging world. Unlike the
Client-Server Architecture, the game makes use of peer to peer
architecture to provide connection between both the players. This game
can also be played by an Xbox 360 controller connected to the PC through
USB. This game was developed using Haaf's Game Engine, which provides
simplified development of Games. This game has a Level Editor, which
allows to create more levels for the game.